+++ apps/web/components/utils.tsx
@@ -1,6 +1,6 @@
 import { fieldConfig } from "@autoform/react";
 import { ZodProvider } from "@autoform/zod";
-import { YupProvider } from "@autoform/yup";
+import { YupProvider, fieldConfig as yupFieldConfig } from "@autoform/yup";
 import * as z from "zod";
 import { object, string, number, date, InferType, array, mixed } from "yup";
 
@@ -13,74 +13,75 @@ enum Sports {
 }
 
 const zodFormSchema = z.object({
-  username: z
-    .string({
-      required_error: "Username is required.",
-    })
-    .min(2, {
-      message: "Username must be at least 2 characters.",
-    })
-    .superRefine(
-      fieldConfig({
-        description: "You cannot change this later.",
-      })
-    ),
+  hobbies: z.array(z.string()).optional(),
+  // username: z
+  //   .string({
+  //     required_error: "Username is required.",
+  //   })
+  //   .min(2, {
+  //     message: "Username must be at least 2 characters.",
+  //   })
+  //   .superRefine(
+  //     fieldConfig({
+  //       description: "You cannot change this later.",
+  //     })
+  //   ),
 
-  password: z
-    .string({
-      required_error: "Password is required.",
-    })
-    .describe("Your secure password")
-    .min(8, {
-      message: "Password must be at least 8 characters.",
-    })
-    .superRefine(
-      fieldConfig({
-        description: (
-          <>
-            Always use a <b>secure password</b>!
-          </>
-        ),
-        inputProps: {
-          type: "password",
-        },
-      })
-    ),
+  // password: z
+  //   .string({
+  //     required_error: "Password is required.",
+  //   })
+  //   .describe("Your secure password")
+  //   .min(8, {
+  //     message: "Password must be at least 8 characters.",
+  //   })
+  //   .superRefine(
+  //     fieldConfig({
+  //       description: (
+  //         <>
+  //           Always use a <b>secure password</b>!
+  //         </>
+  //       ),
+  //       inputProps: {
+  //         type: "password",
+  //       },
+  //     })
+  //   ),
 
-  favouriteNumber: z.coerce
-    .number({
-      invalid_type_error: "Favourite number must be a number.",
-    })
-    .min(1, {
-      message: "Favourite number must be at least 1.",
-    })
-    .max(10, {
-      message: "Favourite number must be at most 10.",
-    })
-    .default(1)
-    .optional(),
+  // favouriteNumber: z.coerce
+  //   .number({
+  //     invalid_type_error: "Favourite number must be a number.",
+  //   })
+  //   .min(1, {
+  //     message: "Favourite number must be at least 1.",
+  //   })
+  //   .max(10, {
+  //     message: "Favourite number must be at most 10.",
+  //   })
+  //   .default(1)
+  //   .optional(),
 
-  acceptTerms: z
-    .boolean()
-    .describe("Accept terms and conditions.")
-    .refine((value) => value, {
-      message: "You must accept the terms and conditions.",
-      path: ["acceptTerms"],
-    }),
+  // acceptTerms: z
+  //   .boolean()
+  //   .describe("Accept terms and conditions.")
+  //   .refine((value) => value, {
+  //     message: "You must accept the terms and conditions.",
+  //     path: ["acceptTerms"],
+  //   }),
 
-  sendMeMails: z.boolean().optional(),
+  // sendMeMails: z.boolean().optional(),
 
-  birthday: z.coerce.date().optional(),
+  // birthday: z.coerce.date().optional(),
 
-  color: z.enum(["red", "green", "blue"]).optional(),
+  // color: z.enum(["red", "green", "blue"]).optional(),
 
-  // Another enum example
-  marshmallows: z
-    .enum(["not many", "a few", "a lot", "too many"])
-    .describe("How many marshmallows fit in your mouth?"),
+  // // Another enum example
+  // marshmallows: z
+  //   .enum(["not many", "a few", "a lot", "too many"])
+  //   .describe("How many marshmallows fit in your mouth?"),
 
-  // Native enum example
-  sports: z.nativeEnum(Sports).describe("What is your favourite sport?"),
+  // // Native enum example
+  // sports: z.nativeEnum(Sports).describe("What is your favourite sport?"),
 
   guests: z.array(
     z.object({
@@ -89,17 +90,27 @@ const zodFormSchema = z.object({
     })
   ),
 
-  location: z.object({
-    city: z.string().optional(),
-    country: z.string().optional(),
-  }),
+  // location: z.object({
+  //   city: z.string().optional(),
+  //   country: z.string().optional(),
+  // }),
 });
 
 export const zodSchemaProvider = new ZodProvider(zodFormSchema);
 
 const yupFormSchema = object({
   name: string().required().label("Your Name").default("John Doe"),
-  age: number().required().positive().integer(),
+  age: number()
+    .required("We need your age to verify you are able to receive discounts.")
+    .positive()
+    .integer(),
+  password: string().transform(
+    yupFieldConfig({
+      inputProps: {
+        type: "password",
+      },
+    })
+  ),
   email: string()
     .email()
     .transform((val) => val),
@@ -111,6 +122,7 @@ const yupFormSchema = object({
     })
   ),
   sport: mixed().oneOf(Object.values(Sports)),
+  hobbies: array().of(string()),
 });
 
 export const yupSchemaProvider = new YupProvider(yupFormSchema);