+++ packages/core/src/index.ts
@@ -1,5 +1,4 @@
 export * from "./types";
-export * from "./dependency-manager";
 export * from "./label";
 export * from "./schema-provider";
 export * from "./logic";