+++ packages/shadcn/src/components/ui/autoform/utils.ts
@@ -1,11 +1,13 @@
 import { FieldConfig } from "@autoform/core";
-import { SuperRefineFunction } from "@autoform/zod";
-import { fieldConfig as baseFieldConfig } from "@autoform/react";
+import {
+  SuperRefineFunction,
+  fieldConfig as baseFieldConfig,
+} from "@autoform/zod";
 import { ReactNode } from "react";
 import { FieldTypes } from "./AutoForm";
 
 export function fieldConfig(
   config: FieldConfig<ReactNode, FieldTypes>
 ): SuperRefineFunction {
-  return baseFieldConfig<FieldTypes>(config);
+  return baseFieldConfig<ReactNode, FieldTypes>(config);
 }