+++ packages/zod/src/default-values.ts
@@ -14,7 +14,7 @@ export function getDefaultValueInZodStack(schema: z.ZodTypeAny): any {
 }
 
 export function getDefaultValues(
-  schema: ZodObjectOrWrapped,
+  schema: ZodObjectOrWrapped
 ): Record<string, any> {
   const objectSchema =
     schema instanceof z.ZodEffects ? schema.innerType() : schema;