+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @autoform/core
 
+## 1.2.0
+
+### Minor Changes
+
+- Improve DX
+
 ## 1.1.3
 
 ### Patch Changes