+++ packages/mui/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @autoform/mui
 
+## 1.2.0
+
+### Minor Changes
+
+- Improve DX
+
+### Patch Changes
+
+- Updated dependencies
+  - @autoform/react@1.3.0
+  - @autoform/core@1.2.0
+  - @autoform/zod@1.2.0
+
 ## 1.1.3
 
 ### Patch Changes