+++ packages/yup/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @autoform/zod
 
+## 1.1.0
+
+### Minor Changes
+
+- Improve DX
+
+### Patch Changes
+
+- Updated dependencies
+  - @autoform/core@1.2.0
+
 ## 1.0.1
 
 ### Patch Changes